수학 u1 변환 NPE 수정 — RunBuilder 폰트크기 삼항 오토언박싱
grepCharStyle 분기(#134 afcffe32 "Fix GREP text styles")가 추가되며
삼항 오토언박싱 NPE 유입:

  Integer x = cond ? (int) foo() : bar();  // bar() null → int 통일 언박싱 → NPE

true 분기가 primitive int 라 Java 가 삼항 전체를 int 로 맞추려 false 분기의
Integer(resolveFontSizeHwpunitsWithConfidence, null 가능)를 언박싱한다. 아래
null 체크는 이미 언박싱 뒤라 무용. GREP math 가 많은 수학 u1 이 이 경로를
타 StoryLoader 초입에서 죽었다(영어 u1 은 grepCharStyle=null 이라 무증상).

(int) → (Integer)(int) 로 박싱해 삼항 전체를 Integer 유지, null 체크 유효화.

검증: 수학 u1 정상 변환(40p, 685수식, warnings=0). 영어 u1 회귀 없음
(20p, warnings=0). false 분기·값 불변.

+++ converter/src/main/java/kr/dogfoot/hwpxlib/tool/idmlconverter/normalizer/resolved/phase3/RunBuilder.java
@@ -186,8 +186,11 @@ class RunBuilder {
             }
             tr.fontFamily(resolvedFontFamily);
         }
+        // 삼항 오토언박싱 주의: true 분기가 primitive int 면 Java 가 양쪽을 int 로
+        // 통일하려 false 분기의 Integer 를 언박싱한다 → null 이면 여기서 NPE.
+        // (int) 대신 (Integer) 로 박싱해 삼항 전체를 Integer 로 유지, 아래 null 체크 유효화.
         Integer resolvedFontSize = grepCharStyle != null && grepCharStyle.fontSize() != null
-                ? (int) CoordinateConverter.pointsToHwpunits(grepCharStyle.fontSize())
+                ? (Integer) (int) CoordinateConverter.pointsToHwpunits(grepCharStyle.fontSize())
                 : RunPropertyResolver.resolveFontSizeHwpunitsWithConfidence(
                         rr, cr, sc.fontSize, confidence);
         if (resolvedFontSize != null) {
